+++ Servico/Service/ServiceIdioma.cs
@@ -92,7 +92,7 @@ namespace Servico.Service
 
 
         //get produto ID
-        public static IdiomaModelView GetIdiomaId(Int16 id)
+        public static IdiomaModelView GetIdiomaId(int id)
         {
             Idioma objretorno = new Idioma();
